- Оптимизация алгоритмов отбора рубрик и терминов

+++ src/main/kotlin/com/gearsy/scitechsearchengine/db/neo4j/repository/projection/RubricHierarchyProjection.kt
@@ -3,6 +3,6 @@ package com.gearsy.scitechsearchengine.db.neo4j.repository.projection
 interface RubricHierarchyProjection {
     fun getParentCipher(): String
     fun getParentTitle(): String
-    fun getParentEmbedding(): List<Float>
+    fun getParentEmbedding(): List<Double>
     fun getChildCiphers(): List<String>
 }
