+++ src/main/kotlin/com/gearsy/scitechsearchengine/db/neo4j/repository/projection/RubricTermProjection.kt
@@ -3,5 +3,5 @@ package com.gearsy.scitechsearchengine.db.neo4j.repository.projection
 interface RubricTermProjection {
     fun getRubricCipher(): String
     fun getContent(): String
-    fun getEmbedding(): List<Float>
+    fun getEmbedding(): List<Double>
 }
